+++ packages/dnb-design-system-portal/src/docs/404.js
@@ -5,7 +5,7 @@
 
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import styled from '@emotion/styled'
+import { GlobalError } from 'dnb-ui-lib/src/components'
 
 export default class PageNotFound extends PureComponent {
   static propTypes = {
@@ -15,27 +15,15 @@ export default class PageNotFound extends PureComponent {
     const { pathname } = this.props.location
 
     return (
-      <Center className="dnb-spacing">
-        <Inner>
-          <h1 className="dnb-h1">404</h1>
-          <h2 className="dnb-h2">We couldn't find that page:</h2>
-          <p className="dnb-p">
+      <GlobalError
+        status="404"
+        title="We couldn't find that page"
+        text={
+          <>
             There's not a page at <code>{pathname}</code>
-          </p>
-        </Inner>
-      </Center>
+          </>
+        }
+      />
     )
   }
 }
-
-const Center = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-`
-const Inner = styled.div`
-  width: 60vw;
-  text-align: center;
-  padding-bottom: 4rem;
-`