+++ packages/dnb-ui-lib/src/components/global-error/style/themes/ui.js
@@ -1,0 +1,6 @@
+/**
+ * Imports the default theme
+ *
+ */
+
+import './dnb-global-error-theme-ui.scss'